+++ src/DotNext.Tests/Net/Cluster/Consensus/Raft/Http/Embedding/EmbeddedClusterSupportTests.cs
@@ -327,6 +327,11 @@ namespace DotNext.Net.Cluster.Consensus.Raft.Http.Embedding
                 await host.StartAsync();
                 object service = host.Services.GetService<ICluster>();
                 NotNull(service);
+                var count = 0;
+                foreach(var member in host.Services.GetService<ICluster>().Members)
+                    if(!member.IsRemote)
+                        count += 1;
+                Equal(1, count);
                 service = host.Services.GetService<IExpandableCluster>();
                 NotNull(service);
                 service = host.Services.GetService<IRaftCluster>();