+++ src/DotNext.Tests/Net/Cluster/Consensus/Raft/Http/Hosting/HostedClusterSupportTests.cs
@@ -14,9 +14,9 @@ namespace DotNext.Net.Cluster.Consensus.Raft.Http.Hosting
             {
                 {"partitioning", "false"},
                 {"metadata:nodeName", "TestNode"},
-                {"port", "3262"},
-                {"members:0", "http://localhost:3262"},
-                {"members:1", "http://localhost:3263"},
+                {"port", "3565"},
+                {"members:0", "http://localhost:3565"},
+                {"members:1", "http://localhost:3566"},
                 {"allowedNetworks:0", "127.0.0.0"}
             };
             using (var host = CreateHost<WebApplicationSetup>(3100, true, config))
@@ -24,6 +24,12 @@ namespace DotNext.Net.Cluster.Consensus.Raft.Http.Hosting
                 await host.StartAsync();
                 object service = host.Services.GetService<ICluster>();
                 NotNull(service);
+                //check whether the local member present
+                var count = 0;
+                foreach (var member in host.Services.GetService<ICluster>().Members)
+                    if (!member.IsRemote)
+                        count += 1;
+                Equal(1, count);
                 service = host.Services.GetService<IExpandableCluster>();
                 NotNull(service);
                 service = host.Services.GetService<IRaftCluster>();