+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Release Notes
 ====
+# 08-21-2019
+<a href="https://www.nuget.org/packages/dotnext.aspnetcore.cluster/0.3.1">DotNext.AspNetCore.Cluster 0.3.1</a>
+* Fixed detection of local IP address
+* Improved IPv6 support
+
 # 08-20-2019
 
 <a href="https://www.nuget.org/packages/dotnext/0.13.0">DotNext 0.13.0</a>